requirements.txt
streamlit==1.31.0
supabase==2.3.0
openai==1.12.0
python-dotenv==1.0.0
numpy==1.24.3
tiktoken==0.5.1
streamlit==1.31.0
supabase==2.3.0
openai==1.12.0
python-dotenv==1.0.0
pdfplumber==0.10.3
PyPDF2==3.0.1
tqdm==4.66.1
pytz==2024.1